refactor: minor changes in project files
