+++ .env
@@ -9,10 +9,11 @@ JWT_SECRET_KEY=secret123123
 JWT_SECRET_REFRESH_KEY=secret123123
 TOKEN_EXPIRE_TIME=1h
 TOKEN_REFRESH_EXPIRE_TIME=24h
+
 DATABASE_HOST=db
 DATABASE_NAME=postgres
 DATABASE_USER=postgres
 DATABASE_PASSWORD=mysecretpassword
-POSTGRES_PORT=5432
+DATABASE_PORT=5432
 DATABASE_DATA=/var/lib/postgresql/data
 DATABASE_VOLUMES=./database/postgres-data/
