+++ development.env
@@ -15,7 +15,7 @@ DATABASE_HOST=db
 DATABASE_NAME=postgres
 DATABASE_USER=postgres
 DATABASE_PASSWORD=mysecretpassword
-POSTGRES_PORT=5432
+DATABASE_PORT=5432
 DATABASE_DATA=/var/lib/postgresql/data
 DATABASE_VOLUMES=./database/postgres-data/
