+++ docker-compose.yml
@@ -22,17 +22,16 @@ services:
       - .env
     build: ${PATH_TO_DB}
     ports:
-      - '${POSTGRES_PORT}:${POSTGRES_PORT}'
+      - ${DATABASE_PORT}:${DATABASE_PORT}
     restart: always
     environment:
-      - POSTGRES_DB=${DATABASE_NAME}
       - PATH_TO_DATABASE=${PATH_TO_DB}
-      - POSTGRES_PORT=${POSTGRES_PORT}
+      - POSTGRES_PORT=${DATABASE_PORT}
       - POSTGRES_USER=${DATABASE_USER}
       - POSTGRES_PASSWORD=${DATABASE_PASSWORD}
       - PGDATA=${DATABASE_DATA}
     volumes:
-      - ${DATABASE_VOLUMES}:${DATABASE_DATA}
+      - ${WORK_DIR}/node_modules
     networks:
         - hl-network
 networks:
