+++ package.json
@@ -18,9 +18,11 @@
     "test:auth": "cross-env TEST_MODE=auth jest --config ./test/jest-e2e.json --noStackTrace --runInBand",
     "test:watch": "jest --watch",
     "test:cov": "jest --coverage",
+    "docker:scan:app": "docker scout cves jupitermass/home-library-service:home-library-service",
+    "docker:scan:db": "docker scout cves jupitermass/home-library-service:home-library-service",
     "test:debug": "node --inspect-brk -r tsconfig-paths/register -r ts-node/register node_modules/.bin/jest --runInBand",
     "typeorm": "npx typeorm -d ./dist/typeorm.config.js",
-    "migration:generate": "npm run build && npm run typeorm -- migration:generate -o",
+    "migration:generate": " npm run typeorm -- migration:generate -o",
     "migration:run": "npm run typeorm migration:run",
     "start:migration:dev": "npm run migration:generate -- ./dist/src/migrations/table && npm run migration:run && npm run start:dev",
     "migration:revert": "npm run typeorm -- migration:revert"
@@ -40,7 +42,6 @@
     "class-validator": "^0.14.0",
     "dotenv": "^16.3.1",
     "http-status-codes": "^2.2.0",
-    "pg": "^8.11.2",
     "reflect-metadata": "^0.1.13",
     "rimraf": "^5.0.1",
     "rxjs": "^7.8.1",
